refactor: redistribute components to specific directories
